# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Internal Task Management API
  description: |
    Internal Task Management API for chat evaluation tasks.
    
    Upload Excel files containing chat evaluation data and get structured evaluation results.
    
    **Core Features:**
    - Upload Excel files with chat evaluation data (questions, golden answers, citations)
    - Task lifecycle management (create, query, cancel, delete)
    - Download evaluation results as Excel files
    
    **Chat Evaluation Requirements:**
    - Excel columns "question", "golden_answer", "golden_citations"
    
    **Authentication:**
    - JWT-based authentication from internal SSO
    - User-scoped access to tasks and operations
  version: 1.0.0
  contact:
    name: Internal API Team
    email: api-team@organization.com
  license:
    name: Internal Use Only
    url: https://internal.organization.com/licenses

servers:
  - url: https://internal-api.organization.com/rest/api/v1
    description: Internal Production API Server
  - url: https://staging-api.organization.com/rest/api/v1
    description: Internal Staging API Server

components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from internal authentication service.
        Token should contain user_id, roles, and permissions.

  schemas:
    # Core Error Schema
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              description: "Error code for programmatic handling"
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: "Human-readable error message"
              example: "Invalid Excel file format"
            details:
              type: string
              description: "Additional error details"
              example: "File must be .xlsx or .xls format"
            timestamp:
              type: string
              format: date-time
              description: "When the error occurred"
              example: "2024-01-01T00:00:00Z"
            request_id:
              type: string
              description: "Request ID for debugging and tracing"
              example: "request_12345"
            user_id:
              type: string
              description: "User ID associated with the request"
              example: "user_123"

    # Pagination Schema
    PaginationMeta:
      type: object
      required:
        - per_page
        - total
        - has_more
      properties:
        per_page:
          type: integer
          minimum: 1
          maximum: 100
          description: "Number of items per page"
          example: 20
        total:
          type: integer
          minimum: 0
          description: "Total number of items available"
          example: 100
        next_cursor:
          type: string
          nullable: true
          description: "Cursor for next page, null if no more results"
          example: "123"
        has_more:
          type: boolean
          description: "Whether there are more results available"
          example: true



    # Enum Schemas
    TaskStatus:
      type: string
      enum: [queueing, processing, completed, cancelled, failed]
      description: |
        Task status values:
        - queueing: Task created with parsed data, waiting for background processing
        - processing: Background processing is executing on the parsed rows
        - completed: All rows processed successfully
        - cancelled: Task cancelled by user
        - failed: Background processing failed with error

    TaskType:
      type: string
      enum: [chat-evaluation]
      description: |
        Task type values:
        - chat-evaluation: Chat evaluation task with questions and expected answers

    # Core Task Schema
    Task: 
      type: object
      required:
        - id
        - user_id
        - original_filename
        - sheet_name
        - task_type
        - task_status
        - upload_batch_id
        - row_count
        - processed_rows
        - created_at
      properties:
        id:
          type: integer
          format: int64
          description: "Unique task identifier"
          example: 12345
        user_id:
          type: string
          description: "ID of the user who created the task"
          example: "user_123"
        original_filename:
          type: string
          description: "Original Excel filename"
          example: "monthly_reports.xlsx"
        sheet_name:
          type: string
          description: "Excel sheet name that was processed"
          example: "January_Data"
        task_type:
          $ref: '#/components/schemas/TaskType'
        task_status:
          $ref: '#/components/schemas/TaskStatus'
        upload_batch_id:
          type: integer
          format: int64
          description: "Groups tasks from the same Excel file upload"
          example: 67890
        row_count:
          type: integer
          minimum: 0
          description: "Total number of rows parsed from Excel sheet"
          example: 150
        processed_rows:
          type: integer
          minimum: 0
          description: "Number of rows completed by background processing"
          example: 75
        error_message:
          type: string
          nullable: true
          description: "Error message if task failed"
          example: "API timeout on row 15"
        metadata:
          type: object
          nullable: true
          description: "Additional task metadata"
          example: {"detected_columns": ["question", "golden_answer"], "extra_columns": ["priority"]}
        created_at:
          type: string
          format: date-time
          description: "When the task was created"
          example: "2024-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: "When the task was last updated"
          example: "2024-01-01T00:05:00Z"
        started_at:
          type: string
          format: date-time
          nullable: true
          description: "When background processing started"
          example: "2024-01-01T00:01:00Z"
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: "When the task was completed"
          example: "2024-01-01T00:10:00Z"
        cancelled_at:
          type: string
          format: date-time
          nullable: true
          description: "When the task was cancelled"







    # Request/Response Schemas

    TaskList:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
          description: "Array of task objects"
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    UploadResponse:
      type: object
      required:
        - upload_batch_id
        - tasks
        - total_sheets
      properties:
        upload_batch_id:
          type: integer
          format: int64
          description: "Unique identifier for this upload batch"
          example: 67890
        tasks:
          type: array
          items:
            type: object
            required:
              - task_id
              - sheet_name
              - task_type
              - status
              - row_count
            properties:
              task_id:
                type: integer
                format: int64
                description: "Unique task identifier"
                example: 12345
              sheet_name:
                type: string
                description: "Excel sheet name"
                example: "Sheet1"
              task_type:
                $ref: '#/components/schemas/TaskType'
              status:
                $ref: '#/components/schemas/TaskStatus'
              row_count:
                type: integer
                minimum: 0
                description: "Number of rows parsed from this sheet"
                example: 50
          description: "Array of created tasks"
        total_sheets:
          type: integer
          minimum: 1
          description: "Total number of sheets processed"
          example: 3
        message:
          type: string
          description: "Success message"
          example: "Successfully created 3 tasks from uploaded Excel file"



  responses:
    # Error Responses (used across all endpoints)
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or expired JWT token"
              details: "Please refresh your authentication token"
              timestamp: "2024-01-15T14:30:00Z"
              request_id: "request_12345"
    
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "FORBIDDEN"
              message: "Access denied to task"
              details: "You can only access your own tasks"
              timestamp: "2024-01-15T14:30:00Z"
              request_id: "request_12345"
              user_id: "user_123"
    
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "Task not found"
              details: "Task with ID 12345 does not exist or you don't have access"
              timestamp: "2024-01-15T14:30:00Z"
              request_id: "request_12345"
              user_id: "user_123"
    
    InternalServerError:
      description: Internal Server Error - Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "INTERNAL_SERVER_ERROR"
              message: "An unexpected error occurred"
              details: "Please contact support if this error persists"
              timestamp: "2024-01-15T14:30:00Z"
              request_id: "request_12345"
              user_id: "user_123"



security:
  - JWTAuth: []

paths:
  /tasks:
    post:
      tags:
        - Tasks
      summary: Upload Excel file and create chat evaluation tasks
      operationId: uploadTasks
      description: |
        Upload an Excel file with chat evaluation data for processing.
        
        **Chat Evaluation Requirements:**
        - **Required columns**: "question", "golden_answer", "golden_citations"
        
        **File Requirements:**
        - Format: .xlsx or .xls
        - Maximum size: 50MB
        - Maximum sheets: 20 per file
        - Maximum rows per sheet: 1,000
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - task_type
              properties:
                file:
                  type: string
                  format: binary
                  description: Excel file (.xlsx or .xls)
                task_type:
                  $ref: '#/components/schemas/TaskType'
                  description: Task type for the uploaded data
                description:
                  type: string
                  maxLength: 500
                  description: Optional description for the upload batch
                  example: "Monthly evaluation data upload"
      responses:
        '201':
          description: Excel file processed and tasks created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Bad Request - File upload validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_file_format:
                  summary: Invalid Excel file format
                  value:
                    error:
                      code: "INVALID_FILE_FORMAT"
                      message: "File must be Excel format (.xlsx or .xls)"
                      details: "Uploaded file has MIME type text/plain, expected Excel format"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                missing_required_columns:
                  summary: Missing required Excel columns
                  value:
                    error:
                      code: "MISSING_REQUIRED_COLUMNS"
                      message: "Excel file missing required columns for chat evaluation"
                      details: "Excel must contain columns: question, golden_answer, golden_citations. Found: question, answer"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                too_many_rows:
                  summary: Too many rows in Excel sheet
                  value:
                    error:
                      code: "TOO_MANY_ROWS"
                      message: "Excel sheet exceeds maximum row limit"
                      details: "Sheet 'Data' contains 1,500 rows. Maximum allowed is 1,000 rows per sheet"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                too_many_sheets:
                  summary: Too many sheets in Excel file
                  value:
                    error:
                      code: "TOO_MANY_SHEETS"
                      message: "Excel file exceeds maximum sheet limit"
                      details: "File contains 25 sheets. Maximum allowed is 20 sheets per file"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                missing_task_type:
                  summary: Missing task_type field
                  value:
                    error:
                      code: "MISSING_REQUIRED_FIELD"
                      message: "task_type is required"
                      details: "task_type field must be provided in the request body"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                invalid_task_type:
                  summary: Invalid task_type value
                  value:
                    error:
                      code: "INVALID_TASK_TYPE"
                      message: "Invalid task_type value"
                      details: "task_type must be 'chat-evaluation'. Received: 'unknown-type'"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                file_too_large:
                  summary: File size exceeds maximum limit
                  value:
                    error:
                      code: "FILE_TOO_LARGE"
                      message: "File size exceeds maximum limit"
                      details: "Maximum file size is 50MB. Uploaded file is 75MB"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: "[NOT FINALIZED] Too Many Requests - Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: "RATE_LIMIT_EXCEEDED"
                  message: "Too many task submissions"
                  details: "🚧 NOT FINALIZED: Each user can only submit 1 task per minute. Please wait 45 seconds before trying again. Rate limiting policy and implementation details need review."
                  timestamp: "2024-01-15T14:30:00Z"
                  request_id: "request_12345"
                  user_id: "user_123"
          headers:
            Retry-After:
              description: "Number of seconds to wait before making another request"
              schema:
                type: integer
                example: 45
            X-RateLimit-Limit:
              description: "Maximum number of requests allowed per minute"
              schema:
                type: integer
                example: 1
            X-RateLimit-Remaining:
              description: "Number of requests remaining in the current window"
              schema:
                type: integer
                example: 0
            X-RateLimit-Reset:
              description: "Unix timestamp when the rate limit window resets"
              schema:
                type: integer
                example: 1705327860
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Tasks
      summary: List user tasks
      description: |
        Get a list of tasks for the authenticated user using pagination.
        
        **Pagination:**
        - First request: Do not include cursor parameter
        - Subsequent requests: Use next_cursor from previous response
        
        **Note:** This endpoint returns task metadata only.
        Use the file download endpoint to get evaluation results.
      operationId: listTasks
      parameters:
        - name: per_page
          in: query
          description: Number of items per page
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: task_type
          in: query
          description: Task type filter
          required: true
          schema:
            $ref: '#/components/schemas/TaskType'
        - name: cursor
          in: query
          description: |
            Cursor for pagination (omit for the first page).
            **Implementation:** The cursor value is the task_id (an incremental integer).
            **Usage:** Use the lowest task_id from the current page as the cursor for the next page.
            **Ordering:** Results are returned newest first (descending by task_id).
          required: false
          schema:
            type: string
            example: "12345"
      responses:
        '200':
          description: List of user's tasks (metadata only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '400':
          description: Bad Request - Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_per_page:
                  summary: Invalid per_page parameter
                  value:
                    error:
                      code: "INVALID_PARAMETER"
                      message: "per_page parameter is invalid"
                      details: "per_page must be between 1 and 100. Received: 150"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                invalid_task_type_filter:
                  summary: Invalid task_type filter
                  value:
                    error:
                      code: "INVALID_TASK_TYPE"
                      message: "Invalid task_type filter"
                      details: "task_type must be 'chat-evaluation'. Received: 'unknown-type'"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                invalid_cursor:
                  summary: Invalid cursor format
                  value:
                    error:
                      code: "INVALID_CURSOR"
                      message: "Invalid cursor format"
                      details: "cursor must be a valid task_id integer. Received: 'abc123'"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Task ID
        schema:
          type: integer
          format: int64
          minimum: 1

    put:
      tags:
        - Tasks
      summary: Cancel task
      operationId: cancelTask
      description: |
        Cancel a task by setting cancelled=true. Only tasks in 'queueing' or 'processing'
        status can be cancelled. Only task owners can cancel their tasks.
      parameters:
        - name: cancelled
          in: query
          description: Must be set to true to cancel the task
          required: true
          schema:
            type: boolean
            enum: [true]
            example: true
      responses:
        '200':
          description: Task cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad Request - Invalid operation for current task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                cannot_cancel:
                  summary: Cannot cancel task in current status
                  value:
                    error:
                      code: "INVALID_STATUS"
                      message: "Cannot cancel completed task"
                      details: "Task has already finished processing and cannot be cancelled"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                invalid_cancelled_parameter:
                  summary: Invalid cancelled parameter
                  value:
                    error:
                      code: "INVALID_PARAMETER"
                      message: "cancelled parameter must be true"
                      details: "To cancel a task, set cancelled=true in query parameters"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Tasks
      summary: Delete task
      operationId: deleteTask
      description: |
        Delete a task permanently. This will remove the task record and 
        associated structured data from input and output tables. 
        Only task owners can delete their tasks.
        Tasks in 'processing' status cannot be deleted.
      responses:
        '204':
          description: Task deleted successfully
        '400':
          description: Bad Request - Invalid operation for current task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: "INVALID_STATUS"
                  message: "Cannot delete task in processing status"
                  details: "Cancel the task first, then delete it"
                  timestamp: "2024-01-15T14:30:00Z"
                  request_id: "request_12345"
                  user_id: "user_123"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /tasks/{id}/file:
    parameters:
      - name: id
        in: path
        required: true
        description: Task ID
        schema:
          type: integer
          format: int64
          minimum: 1

    get:
      tags:
        - Tasks
      summary: Download task evaluation results Excel file
      operationId: downloadTaskResults
      description: |
        Download the complete evaluation results as a single Excel file (.xlsx).
        Only task owners can download their results.
        
        **File Contents (4 sheets):**
        - **QnAs**: Original questions and golden answers from uploaded data
        - **API_Responses**: Generated answers and citations from the evaluation system
        - **Answer_Evaluation**: Similarity scores and metrics for answer comparison
        - **Citation_Evaluation**: Citation relevance and accuracy analysis
        
        **Requirements:**
        - Task must be in 'completed' status
        - User must be the task owner
        - Results must be available (background processing finished successfully)
        
        **File Naming:** `task_{id}_evaluation_results.xlsx`
      responses:
        '200':
          description: Successfully generated and returned Excel file with evaluation results
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
                description: "Excel (.xlsx) file containing complete evaluation results across 4 sheets"
          headers:
            Content-Disposition:
              description: "Attachment header with generated filename"
              schema:
                type: string
                example: 'attachment; filename="task_12345_evaluation_results.xlsx"'
            Content-Length:
              description: "Size of the Excel file in bytes"
              schema:
                type: integer
                minimum: 1
                example: 245760
            Content-Type:
              description: "MIME type for Excel files"
              schema:
                type: string
                example: "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
            Last-Modified:
              description: "When the results file was last generated"
              schema:
                type: string
                example: "Wed, 15 Jan 2024 14:30:00 GMT"
        '400':
          description: Bad Request - Cannot download results for current task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                task_not_completed:
                  summary: Task processing not completed
                  value:
                    error:
                      code: "TASK_NOT_COMPLETED"
                      message: "Evaluation results not available for download"
                      details: "Task is still processing. Current status: processing (75/150 rows completed)"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                task_failed:
                  summary: Task processing failed
                  value:
                    error:
                      code: "TASK_FAILED"
                      message: "Cannot download results for failed task"
                      details: "Task processing failed. Error: API timeout on row 42"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
                task_cancelled:
                  summary: Task was cancelled
                  value:
                    error:
                      code: "TASK_CANCELLED"
                      message: "Cannot download results for cancelled task"
                      details: "Task was cancelled before completion"
                      timestamp: "2024-01-15T14:30:00Z"
                      request_id: "request_12345"
                      user_id: "user_123"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /task/types:
    get:
      tags:
        - Tasks
      summary: Get available task types
      operationId: getTaskTypes
      description: |
        Get a list of all available and enabled task types.
        This endpoint allows clients to discover what task types are supported.
        
        **Use Cases:**
        - Frontend applications discovering available task types
        - Validation of task type parameters before upload
      responses:
        '200':
          description: Successfully retrieved available task types
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                description: Array of available task type identifiers
                example: ["chat-evaluation", "url-cleaning"]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError' 